@@ -2,11 +2,11 @@ package de.gessnerfl.fakesmtp.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+
 import java.util.*;
 
 import static java.util.Comparator.comparing;
-import static java.util.stream.Collectors.toList;
 
 @Entity
 @Table(name = "email")
@@ -16,36 +16,36 @@ public class Email {
     @GeneratedValue(generator = "email_generator")
     private Long id;
 
-    @Column(name="from_address", length = 255, nullable = false)
+    @Column(name = "from_address", length = 255, nullable = false)
     @Basic(optional = false)
     private String fromAddress;
 
-    @Column(name="to_address", length = 255, nullable = false)
+    @Column(name = "to_address", length = 255, nullable = false)
     @Basic(optional = false)
     private String toAddress;
 
     @Lob
-    @Column(name="subject", nullable = false)
+    @Column(name = "subject", nullable = false)
     @Basic(optional = false)
     private String subject;
 
-    @Column(name="received_on", nullable = false)
+    @Column(name = "received_on", nullable = false)
     @Basic(optional = false)
     @Temporal(TemporalType.TIMESTAMP)
     private Date receivedOn;
 
     @Lob
-    @Column(name="raw_data", nullable = false)
+    @Column(name = "raw_data", nullable = false)
     @Basic(optional = false)
     private String rawData;
 
-    @OneToMany(mappedBy="email", cascade = CascadeType.ALL, orphanRemoval=true)
+    @OneToMany(mappedBy = "email", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<EmailContent> contents = new ArrayList<>();
 
-    @OneToMany(mappedBy="email", cascade = CascadeType.ALL, orphanRemoval=true)
+    @OneToMany(mappedBy = "email", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<EmailAttachment> attachments = new ArrayList<>();
 
-    @OneToMany(mappedBy="email", cascade = CascadeType.ALL, orphanRemoval=true)
+    @OneToMany(mappedBy = "email", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<InlineImage> inlineImages = new ArrayList<>();
 
     public Long getId() {
@@ -88,7 +88,7 @@ public class Email {
         this.receivedOn = receivedOn;
     }
 
-    public void setRawData(String rawData){
+    public void setRawData(String rawData) {
         this.rawData = rawData;
     }
 
@@ -102,20 +102,20 @@ public class Email {
     }
 
     public List<EmailContent> getContents() {
-        return contents.stream().sorted(comparing(EmailContent::getContentType)).collect(toList());
+        return contents.stream().sorted(comparing(EmailContent::getContentType)).toList();
     }
 
     @JsonIgnore
-    public Optional<EmailContent> getPlainContent(){
+    public Optional<EmailContent> getPlainContent() {
         return getContentOfType(ContentType.PLAIN);
     }
 
     @JsonIgnore
-    public Optional<EmailContent> getHtmlContent(){
+    public Optional<EmailContent> getHtmlContent() {
         return getContentOfType(ContentType.HTML);
     }
 
-    private Optional<EmailContent> getContentOfType(ContentType contentType){
+    private Optional<EmailContent> getContentOfType(ContentType contentType) {
         return contents.stream().filter(c -> contentType.equals(c.getContentType())).findFirst();
     }
 
@@ -129,13 +129,13 @@ public class Email {
     }
 
 
-    public void addInlineImage(InlineImage inlineImage){
+    public void addInlineImage(InlineImage inlineImage) {
         inlineImage.setEmail(this);
         this.inlineImages.add(inlineImage);
     }
 
     @JsonIgnore
-    public Optional<InlineImage> getInlineImageByContentId(String cid){
+    public Optional<InlineImage> getInlineImageByContentId(String cid) {
         return inlineImages.stream().filter(i -> i.getContentId().equals(cid)).findFirst();
     }
 
