@@ -3,7 +3,7 @@ package de.gessnerfl.fakesmtp.model;
 import java.util.Locale;
 
 public enum ContentType {
-    HTML, PLAIN, MULTIPART_ALTERNATIVE, MULTIPART_MIXED, MULTIPART_RELATED, UNDEFINED, IMAGE;
+    HTML, PLAIN, MULTIPART_ALTERNATIVE, MULTIPART_MIXED, MULTIPART_RELATED, UNDEFINED, IMAGE, OCTET_STREAM;
 
     public static ContentType fromString(String string) {
         string = string.toLowerCase(Locale.ENGLISH);
@@ -13,6 +13,7 @@ public enum ContentType {
         if (string.startsWith("text/html")) return HTML;
         if (string.startsWith("text/plain")) return PLAIN;
         if (string.startsWith("image/")) return IMAGE;
+        if (string.startsWith("application/octet-stream")) return OCTET_STREAM;
         return UNDEFINED;
     }
 }
